// backend/model/ekyc/ekycproductcode.ts
export enum EKYCProductCode {
    ID_OCR = "ID_OCR",
    EKYC = "eKYC",
    FACE_VERIFY = "FACE_VERIFY",
    FACE_LIVENESS = "FACE_LIVENESS",
}

export enum DocType {
    Passport = "01000000",//Passport for all countries
    HKID2003 = "00000006",//HKID 2003 version (Hong Kong Identity Card)
    HKID2018 = "00000008",//HKID 2018 version (Hong Kong Identity Card)
    ChinaExitEntryPermit = "00000007",//China Exit-Entry Permit for Travelling to and from Hong Kong and Macao
    ChinaMainlandTravelPermit = "00000009",//China Mainland Travel Permit for Hong Kong and Macao Residents
    MacaoID = "000000011",//Macao ID
    ChinaMainlandTravelPermitForTaiwanResidents = "00000012",//China Mainland Travel Permit for Taiwan Residents
}